pyproject.toml
[tool.poetry]
name = "yulita"
version = "0.1.0"
description = ""
authors = ["kalilovm <marlenkalilov86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"
python-decouple = "^3.7"
drf-spectacular = "^0.25.1"
drf-spectacular-sidecar = "^2022.12.1"
pytest = "^7.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
